webdav_server: always use "connection: close" semantics
The Mac OS X webdav mounter doesn't handle keep-alive well.
In particular, user programs that open() a file without
calling write() for a while will fail a subsequent write().
Arguably this is the fault of buggy user program but failing
write() is a rare occurence. To avoid a spuriously failing write()
the webdav mounter should reopen the connection when a request
failes but it doesn't. This change makes it so the webdav mounter
doesn't attempt to reuse connections in the first place.